improve seeds

+++ db/seeds.rb
@@ -1,14 +1,16 @@
+return unless Rails.env.development?
+
 Message.delete_all
 Inbox.delete_all
 User.delete_all
 
 rand(2..3).times do
-  faker_email = Faker::Internet.email
+  faker_email = Faker::Internet.unique.email
   user = User.create(email: faker_email,
                      password: Devise.friendly_token[0, 20])
 
   rand(2..5).times do
-    faker_name = Faker::Quote.famous_last_words
+    faker_name = Faker::Lorem.unique.question(word_count: 5)
     inbox = Inbox.create(name: faker_name,
                  user: user)
 
@@ -16,8 +18,7 @@ rand(2..3).times do
       faker_body = Faker::Lorem.question
       Message.create(body: faker_body,
                      inbox: inbox,
-                     user: user)
-                     # TODO: make it a random user
+                     user: User.all.sample)
                      # add random votes
     end
   end
